+++ src/panoptes/utils/config/__init__.py
@@ -1,0 +1,13 @@
+from panoptes.utils.config.helpers import load_config, save_config
+from panoptes.utils.config.models import DatabaseConfig, DirectoriesConfig, LocationConfig, PANOPTESBaseConfig
+from panoptes.utils.config.watcher import ConfigWatcher
+
+__all__ = [
+    "load_config",
+    "save_config",
+    "DatabaseConfig",
+    "DirectoriesConfig",
+    "LocationConfig",
+    "PANOPTESBaseConfig",
+    "ConfigWatcher",
+]